app/requirements.txt
streamlit==1.30.0
psycopg2-binary
pandas
python-dotenv
plotly
altair<5
